instagram regulation added
