+++ Diary_app/package.json
@@ -34,8 +34,10 @@
     "react-native-picker-select": "^8.0.4",
     "react-native-safe-area-context": "^4.5.0",
     "react-native-screens": "^3.20.0",
+    "react-native-simple-toast": "^2.0.0",
     "react-native-sound": "^0.11.2",
     "react-native-sound-player": "^0.13.2",
+    "react-native-toast-message": "^2.1.6",
     "react-native-vector-icons": "^9.2.0",
     "react-navigation": "^4.4.4",
     "sox": "^0.1.0"
