+++ Diary_app/src/screens/IntroScreen/IntroScreen.tsx
@@ -112,7 +112,7 @@ const IntroScreen = ({ navigation, route }: any) => {
         // clearTimeout(timeout1);
     }, [text]);
 
-
+    const index:number = 0
     return (
         <ImageBackground>
             <IntroImage>
@@ -125,7 +125,9 @@ const IntroScreen = ({ navigation, route }: any) => {
             <CustomButton
                 style={{ position: 'absolute', bottom: '10 %', justifySelf: 'center', alignSelf: 'center' }}
                 title="START"
-                onPress={() => navigation.navigate('TagRecording')}
+                onPress={() => {
+                    navigation.navigate('TagRecording' , {index:index})
+                }}
                 backgroundColor="transparent"
                 textStyle={{ fontSize: 35, color: 'white', fontFamily: 'Zumattan' }} />
